Added all logic into Lesson3.cs Class.
Program.cs calls the Lesson3.Demo function.
